+++ src/Chunkyard.Tests/Infrastructure/DummyPrompt.cs
@@ -4,14 +4,21 @@ namespace Chunkyard.Tests.Infrastructure
 {
     internal class DummyPrompt : IPrompt
     {
+        private readonly string _password;
+
+        public DummyPrompt(string password)
+        {
+            _password = password;
+        }
+
         public string NewPassword()
         {
-            return "secret";
+            return _password;
         }
 
         public string ExistingPassword()
         {
-            return "secret";
+            return _password;
         }
     }
 }